backend/requirements.txt
Django>=5.0
psycopg2-binary
python-dotenv
Pillow
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-filter
django-environ
pytest
pytest-django
stripe
stripe
